filters in climb log now follow formatting of photos.js, WIP: update determineArea helper function
